# Pack multiple vertex and index arrays together into growable buffers.