--- a/crates/bevy_pbr/src/volumetric_fog/render.rs
+++ b/crates/bevy_pbr/src/volumetric_fog/render.rs
@@ -18,7 +18,9 @@ use bevy_ecs::{
 };
 use bevy_math::{vec4, Mat3A, Mat4, Vec3, Vec3A, Vec4, Vec4Swizzles as _};
 use bevy_render::{
-    mesh::{GpuBufferInfo, GpuMesh, Mesh, MeshVertexBufferLayoutRef},
+    mesh::{
+        allocator::MeshAllocator, Mesh, MeshVertexBufferLayoutRef, RenderMesh, RenderMeshBufferInfo,
+    },
     render_asset::RenderAssets,
     render_graph::{NodeRunError, RenderGraphContext, ViewNode},
     render_resource::{
@@ -329,6 +331,7 @@ impl ViewNode for VolumetricFogNode {
         let volumetric_lighting_uniform_buffers = world.resource::<VolumetricFogUniformBuffer>();
         let image_assets = world.resource::<RenderAssets<GpuImage>>();
         let msaa = world.resource::<Msaa>();
+        let mesh_allocator = world.resource::<MeshAllocator>();
 
         // Fetch the uniform buffer and binding.
         let (
@@ -344,7 +347,7 @@ impl ViewNode for VolumetricFogNode {
             return Ok(());
         };
 
-        let gpu_meshes = world.resource::<RenderAssets<GpuMesh>>();
+        let render_meshes = world.resource::<RenderAssets<RenderMesh>>();
 
         for view_fog_volume in view_fog_volumes.iter() {
             // If the camera is outside the fog volume, pick the cube mesh;
@@ -354,6 +357,11 @@ impl ViewNode for VolumetricFogNode {
                 CUBE_MESH.clone()
             } else {
                 PLANE_MESH.clone()
+            };
+
+            let Some(vertex_buffer_slice) = mesh_allocator.mesh_vertex_slice(&mesh_handle.id())
+            else {
+                continue;
             };
 
             let density_image = view_fog_volume
@@ -370,7 +378,7 @@ impl ViewNode for VolumetricFogNode {
 
             // This should always succeed, but if the asset was unloaded don't
             // panic.
-            let Some(gpu_mesh) = gpu_meshes.get(&mesh_handle) else {
+            let Some(render_mesh) = render_meshes.get(&mesh_handle) else {
                 return Ok(());
             };
 
@@ -426,7 +434,7 @@ impl ViewNode for VolumetricFogNode {
                 .command_encoder()
                 .begin_render_pass(&render_pass_descriptor);
 
-            render_pass.set_vertex_buffer(0, *gpu_mesh.vertex_buffer.slice(..));
+            render_pass.set_vertex_buffer(0, *vertex_buffer_slice.buffer.slice(..));
             render_pass.set_pipeline(pipeline);
             render_pass.set_bind_group(
                 0,
@@ -446,17 +454,23 @@ impl ViewNode for VolumetricFogNode {
             );
 
             // Draw elements or arrays, as appropriate.
-            match &gpu_mesh.buffer_info {
-                GpuBufferInfo::Indexed {
-                    buffer,
+            match &render_mesh.buffer_info {
+                RenderMeshBufferInfo::Indexed {
                     index_format,
                     count,
                 } => {
-                    render_pass.set_index_buffer(*buffer.slice(..), *index_format);
+                    let Some(index_buffer_slice) =
+                        mesh_allocator.mesh_index_slice(&mesh_handle.id())
+                    else {
+                        continue;
+                    };
+
+                    render_pass
+                        .set_index_buffer(*index_buffer_slice.buffer.slice(..), *index_format);
                     render_pass.draw_indexed(0..*count, 0, 0..1);
                 }
-                GpuBufferInfo::NonIndexed => {
-                    render_pass.draw(0..gpu_mesh.vertex_count, 0..1);
+                RenderMeshBufferInfo::NonIndexed => {
+                    render_pass.draw(0..render_mesh.vertex_count, 0..1);
                 }
             }
         }
@@ -584,7 +598,7 @@ pub fn prepare_volumetric_fog_pipelines(
         With<VolumetricFogSettings>,
     >,
     msaa: Res<Msaa>,
-    meshes: Res<RenderAssets<GpuMesh>>,
+    meshes: Res<RenderAssets<RenderMesh>>,
 ) {
     let plane_mesh = meshes.get(&PLANE_MESH).expect("Plane mesh not found!");