--- a/crates/bevy_render/src/batching/gpu_preprocessing.rs
+++ b/crates/bevy_render/src/batching/gpu_preprocessing.rs
@@ -185,7 +185,8 @@ pub struct IndirectParameters {
     /// This field is in the same place in both structures.
     pub instance_count: u32,
 
-    /// The index of the first vertex we're to draw.
+    /// For `ArrayIndirectParameters`, `first_vertex`; for
+    /// `ElementIndirectParameters`, `first_index`.
     pub first_vertex_or_first_index: u32,
 
     /// For `ArrayIndirectParameters`, `first_instance`; for