… crates/bevy_pbr/src/material.rs
… crates/bevy_render/Cargo.toml
… crates/bevy_render/src/mesh/mesh/mod.rs
… crates/bevy_render/src/mesh/mod.rs
… crates/bevy_render/src/render_asset.rs
… crates/bevy_sprite/src/mesh2d/material.rs
… crates/bevy_sprite/src/mesh2d/mesh.rs
… examples/2d/mesh2d_manual.rs
… examples/shader/shader_instancing.rs